✨ history data 가져오기 버튼
- getNewHistory 함수 -> useNewHistory 커스텀 훅 변경
- 히스토리 창 스크롤 제거
- flex shrink 0 설정
- IHistoryData streamId 추가
