+++ client/src/img/index.ts
@@ -20,6 +20,7 @@ import sirenGreenPath from './siren-green.svg';
 import sirenOrangePath from './siren-orange.svg';
 import sirenRedPath from './siren-red.svg';
 import pauseBtnPath from './pauseBtn.svg';
+import primaryPlusCirclePath from './primaryPlusCircle.svg';
 
 export const share: string = sharePath;
 export const logo: string = logoPath;
@@ -43,3 +44,4 @@ export const sirenGreenIcon: string = sirenGreenPath;
 export const sirenOrangeIcon: string = sirenOrangePath;
 export const sirenRedIcon: string = sirenRedPath;
 export const pauseBtn: string = pauseBtnPath;
+export const primaryPlusCircle: string = primaryPlusCirclePath;
