+++ client/src/types/history.ts
@@ -9,4 +9,5 @@ export interface IHistoryData {
   data: THistoryEventData;
   newNodeId?: number;
   historyId: number;
+  streamId: string;
 }
